Split Component and Hardware Infos into two files
